// src/app/modules/detail-fund-management/interfaces/DetailFund.enum.ts
export enum DetailFundFields
{
    PROJECT_FUNDS_COUNT = "ProjectFundsCount",  // số lượng dự án của quỹ
    EMISSARY_COUNT = "ProjectFundsWithUserIdCount",  // số lượng sứ giả của quỹ
    TOTAL_MOMO_AMOUNT = "TotalMomoAmount",  // số tiền donate của quỹ
    TOTAL_USER_DONATE = "TotalSupportersCount",  // số người donate
    USERS = "Users"  // list user của quỹ
}
export enum DetailFundUserFields
{
    AVATAR = "Avatar",  // ẢNH SỨ GIẢ
    FULL_NAME = "FullName",  // TÊN SỨ GIẢ
    TOTAL_DONATE = "TotalDonate",  // số tiền donate của SỨ GIẢ
    
}